Moved destination-adding into a transformer
